SDL: Split the GraphicsManager into Surface and OpenGL
