+++ backends/platform/sdl/sdl.h
@@ -82,6 +82,10 @@ public:
 	// ResidualVM specific code
 	virtual void suggestSideTextures(Graphics::Surface *left,
 	                                 Graphics::Surface *right) override;
+
+	// ResidualVM specific code
+	virtual void setupScreen(uint screenW, uint screenH, bool fullscreen, bool accel3d) override;
+
 protected:
 	bool _inited;
 	bool _initedSDL;
